infrastructure/jobs/send_reminders/requirements.txt
google-cloud-firestore==2.*
google-cloud-pubsub==2.*
